refactor(test): migrate runs.ts mutation/query helpers to three-tier architecture
Move 14 DB-direct seeder functions to db-test-seeders/runs.ts (createCliRun,
markRunningRunsAsCompleted, setTestRunStatus, setTestRunModelProvider,
setTestRunSelectedModel, insertTestZeroRun, expireQueueEntry,
insertTestQueueEntry, createTestCallback, linkRunToSchedule,
insertTestConversation, insertTestSandboxTelemetry, insertTestUsageDaily)
and 9 read-only assertion functions to new db-test-assertions/runs.ts.

api-test-helpers/runs.ts now contains only API route wrappers (createTestRun,
getTestRun, completeTestRun, failTestRun, enqueueTestRun) with re-exports
for backward compatibility. Zero globalThis.services.db calls, zero
db/schema imports, zero initServices calls remain in api-test-helpers.

Closes #9375
